{{cookiecutter.project_slug}}/backend/requirements/local.txt
-r common.txt
psycopg2-binary

# [dev]
autopep8
ipython

# [test]
pytest-django
